Fix uint32 total overflow in computeSingleProb

+++ token.go
@@ -1013,14 +1013,14 @@ func (h *CoeffHistogram) computeProbsForBand(baseProbs, updated *[4][8][3][11]ui
 
 // computeSingleProb computes a single updated probability from histogram counts.
 func (h *CoeffHistogram) computeSingleProb(baseProb uint8, i, j, k, l int) uint8 {
-	falseCount := h.counts[i][j][k][l][0]
-	trueCount := h.counts[i][j][k][l][1]
+	falseCount := uint64(h.counts[i][j][k][l][0])
+	trueCount := uint64(h.counts[i][j][k][l][1])
 	total := falseCount + trueCount
 
 	if total < 16 {
 		return baseProb
 	}
-	return clampProb(uint32(uint64(falseCount) * 256 / uint64(total)))
+	return clampProb(uint32(falseCount * 256 / total))
 }
 
 // clampProb clamps a probability value to the valid range [1, 255].
